Reworked chesspiece movement
Everything i made yesterday was useless. Instead of using XR interaction toolkit's object movement features, we made our own. The reason for this is that the XR interaction toolkit only allows objects to move on detected AR planes (which i cant seem to add to our chessboard prefab), but our new script uses raycasting from the camera to detect gameobjects in world space, and then use the coordinates of the raycasthit to transform the chesspiece position