Place chessboard on plane
Inserted my own picture, added an AR Chessboard Manager that checks if the chessboard has been placed, if it has not been placed then the user will be able to place it on the planes detected. I also changed the near clipping from 0.1 to 0.005 to be able to get closer to the chessboard. Scaled size of chessboard prefab from 0.5 to 1.